refactor : 레이싱카 이름(RacingCarName) package move
